usage-service/requirements.txt
# Web/API
# Bumped to 0.128.6 to satisfy Cadwyn 6.2.0
fastapi==0.128.6
uvicorn[standard]==0.31.1

# Kafka consumer
confluent-kafka==2.6.0
fastavro==1.9.7

# DB migrations
alembic==1.13.3
psycopg2-binary==2.9.10

# Data processing / MLOps
pandas==2.2.3
apache-airflow==2.10.2
apache-airflow-providers-standard==1.3.0
mlflow==2.17.0

# LLM / RAG 
langchain==0.3.4
openai==1.52.0
chromadb==0.5.15
langchain-core==0.3.12
langchain-community==0.3.3
langchain-openai==0.2.3
sentence-transformers==3.2.1

# Utilities & Resilience
python-dotenv==1.0.1
requests==2.32.3
pybreaker==1.4.0
tenacity==9.0.0

# Tracing
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-instrumentation-requests==0.48b0
opentelemetry-instrumentation-sqlalchemy==0.48b0
opentelemetry-instrumentation-logging==0.48b0
